# Implemented_Midterm_Program/openapi.yaml
openapi: 3.0.3

info:
  title: Course Task Tracker API
  version: 1.0.0

servers:
  - url: http://localhost:3000

paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /api/tasks:
    get:
      summary: Get all tasks
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTask"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"

  /api/tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Task id
        schema:
          type: string

    get:
      summary: Get one task by id
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Replace an existing task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTask"
      responses:
        "200":
          description: Task replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Partially update an existing task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTask"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete a task
      responses:
        "204":
          description: Task deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"

components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - course
        - completed
      properties:
        id:
          type: string
          example: "1"
        title:
          type: string
          example: Study for Midterm
        course:
          type: string
          example: CS453
        completed:
          type: boolean
          example: false

    NewTask:
      type: object
      required:
        - title
        - course
        - completed
      properties:
        title:
          type: string
          example: Study for Midterm
        course:
          type: string
          example: CS453
        completed:
          type: boolean
          example: false

    UpdateTask:
      type: object
      properties:
        title:
          type: string
          example: Finish Midterm
        course:
          type: string
          example: CS453
        completed:
          type: boolean
          example: true

    Error:
      type: object
      properties:
        error:
          type: string
          example: Task not found

  responses:
    BadRequest:
      description: Bad request because the client sent invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    NotFound:
      description: Task was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"